cpu/measurement/devices: fix narrowing, TSC divide-by-zero, and underflows
- abstract_cpu: use std::stoi (not stoull) for line_level assignment
- cpu: add cpunr < 0 lower-bound guard in handle_trace_point
- intel_cpus: initialize ratio=0.0 and guard TSC division by zero at
  all three nhm measurement_end sites; prevents Inf/NaN on hypervisors
- extech: change decode_extech_value return type to int (was unsigned int
  returning -1, which wrapped to UINT_MAX); fix parse_packet inner loop
  to iterate 4 times instead of 1
- alsa: fix uint64_t underflow in end_measurement and utilization by
  clamping to 0 when end < start; fix static index sentinel from 0 to -1
  so index==0 is not mistaken for 'not yet initialized'
- devfreq: guard active_time overflow in process_time_stamps; clamp idle
  time to 0 if active_time exceeds sample_time

+++ src/cpu/abstract_cpu.cpp
@@ -217,7 +217,7 @@ void abstract_cpu::insert_cstate(const std::string &linux_name, const std::strin
 			char c = human_name[i];
 			if (c >= '0' && c <= '9') {
 				try {
-					state->line_level = std::stoull(human_name.substr(i));
+					state->line_level = std::stoi(human_name.substr(i));
 				} catch (...) {}
 
 				if (i + 1 < human_name.length() && human_name[i+1] != '-') {
@@ -244,7 +244,7 @@ void abstract_cpu::insert_cstate(const std::string &linux_name, const std::strin
 			char c = linux_name[i];
 			if (c >= '0' && c <= '9') {
 				try {
-					state->line_level = std::stoull(linux_name.substr(i));
+					state->line_level = std::stoi(linux_name.substr(i));
 				} catch (...) {}
 				break;
 			}
